fix: Rename tracer file
