+++ policyengine_household_api/utils/__init__.py
@@ -1,3 +1,3 @@
 from .json import *
-from .tracer import *
+from .computation_tree import *
 from .google_cloud import *
